socket.io-main/requirement.txt
python-socketio==5.4.0
websocket-client==1.2.1
paramiko==2.7.2